Split long files into submodules to make it easier to grok and extend the project
