+++ AUTHORS
@@ -9,3 +9,4 @@ Luke Petre
 Filip Wasilewski (en@ig.ma)
 Sergey Fursov <geyser85@gmail.com>
 Dylann Cordel <cordel.d@free.fr>
+James Keys <skolsuper@gmail.com>
