+++ pybb/forms/__init__.py
@@ -1,0 +1,9 @@
+from .attachment import AttachmentForm, AttachmentFormSet
+from .misc import UserSearchForm
+from .polls import (PollForm, PollAnswerForm, BasePollAnswerFormset,
+                    PollAnswerFormSet)
+from .post import PostForm, AdminPostForm
+try:
+    from .profile import EditProfileForm
+except ImportError:
+    pass
