+++ pybb/forms/post.py
@@ -1,62 +1,24 @@
 # -*- coding: utf-8 -*-
-
 from __future__ import unicode_literals
-import re
+
 import inspect
 
 from django import forms
-from django.core.exceptions import FieldError
-from django.forms.models import inlineformset_factory, BaseInlineFormSet
-from django.utils.translation import ugettext, ugettext_lazy
 from django.utils.timezone import now as tznow
+from django.utils.translation import ugettext, ugettext_lazy as _
 
 from pybb import compat, defaults, util
-from pybb.models import Topic, Post, Attachment, PollAnswer
-
+from pybb.models import Post, Topic
 
 User = compat.get_user_model()
 username_field = compat.get_username_field()
 
 
-class AttachmentForm(forms.ModelForm):
-    class Meta(object):
-        model = Attachment
-        fields = ('file', )
-
-    def clean_file(self):
-        if self.cleaned_data['file'].size > defaults.PYBB_ATTACHMENT_SIZE_LIMIT:
-            raise forms.ValidationError(ugettext('Attachment is too big'))
-        return self.cleaned_data['file']
-
-AttachmentFormSet = inlineformset_factory(Post, Attachment, extra=1, form=AttachmentForm)
-
-
-class PollAnswerForm(forms.ModelForm):
-    class Meta:
-        model = PollAnswer
-        fields = ('text', )
-
-
-class BasePollAnswerFormset(BaseInlineFormSet):
-    def clean(self):
-        forms_cnt = (len(self.initial_forms) + len([form for form in self.extra_forms if form.has_changed()]) -
-                     len(self.deleted_forms))
-        if forms_cnt > defaults.PYBB_POLL_MAX_ANSWERS:
-            raise forms.ValidationError(
-                ugettext('You can''t add more than %s answers for poll' % defaults.PYBB_POLL_MAX_ANSWERS))
-        if forms_cnt < 2:
-            raise forms.ValidationError(ugettext('Add two or more answers for this poll'))
-
-
-PollAnswerFormSet = inlineformset_factory(Topic, PollAnswer, extra=2, max_num=defaults.PYBB_POLL_MAX_ANSWERS,
-                                          form=PollAnswerForm, formset=BasePollAnswerFormset)
-
-
 class PostForm(forms.ModelForm):
-    name = forms.CharField(label=ugettext_lazy('Subject'))
-    poll_type = forms.TypedChoiceField(label=ugettext_lazy('Poll type'), choices=Topic.POLL_TYPE_CHOICES, coerce=int)
+    name = forms.CharField(label=_('Subject'))
+    poll_type = forms.TypedChoiceField(label=_('Poll type'), choices=Topic.POLL_TYPE_CHOICES, coerce=int)
     poll_question = forms.CharField(
-        label=ugettext_lazy('Poll question'),
+        label=_('Poll question'),
         required=False,
         widget=forms.Textarea(attrs={'class': 'no-markitup'}))
 
@@ -164,7 +126,7 @@ class AdminPostForm(PostForm):
     Superusers can post messages from any user and from any time
     If no user with specified name - new user will be created
     """
-    login = forms.CharField(label=ugettext_lazy('User'))
+    login = forms.CharField(label=_('User'))
 
     def __init__(self, *args, **kwargs):
         if args:
@@ -186,65 +148,3 @@ class AdminPostForm(PostForm):
                                'is_staff': False}
             self.user = User.objects.create(**create_data)
         return super(AdminPostForm, self).save(*args, **kwargs)
-
-
-try:
-    class EditProfileForm(forms.ModelForm):
-        class Meta(object):
-            model = util.get_pybb_profile_model()
-            fields = ['signature', 'time_zone', 'language', 'show_signatures', 'avatar']
-
-        def __init__(self, *args, **kwargs):
-            super(EditProfileForm, self).__init__(*args, **kwargs)
-            self.fields['signature'].widget = forms.Textarea(attrs={'rows': 2, 'cols:': 60})
-
-        def clean_avatar(self):
-            if self.cleaned_data['avatar'] and (self.cleaned_data['avatar'].size > defaults.PYBB_MAX_AVATAR_SIZE):
-                forms.ValidationError(ugettext('Avatar is too large, max size: %s bytes' %
-                                               defaults.PYBB_MAX_AVATAR_SIZE))
-            return self.cleaned_data['avatar']
-
-        def clean_signature(self):
-            value = self.cleaned_data['signature'].strip()
-            if len(re.findall(r'\n', value)) > defaults.PYBB_SIGNATURE_MAX_LINES:
-                raise forms.ValidationError('Number of lines is limited to %d' % defaults.PYBB_SIGNATURE_MAX_LINES)
-            if len(value) > defaults.PYBB_SIGNATURE_MAX_LENGTH:
-                raise forms.ValidationError('Length of signature is limited to %d' % defaults.PYBB_SIGNATURE_MAX_LENGTH)
-            return value
-except FieldError:
-    pass
-
-
-class UserSearchForm(forms.Form):
-    query = forms.CharField(required=False, label='')
-
-    def filter(self, qs):
-        if self.is_valid():
-            query = self.cleaned_data['query']
-            return qs.filter(**{'%s__contains' % username_field: query})
-        else:
-            return qs
-
-
-class PollForm(forms.Form):
-    def __init__(self, topic, *args, **kwargs):
-        self.topic = topic
-
-        super(PollForm, self).__init__(*args, **kwargs)
-
-        qs = PollAnswer.objects.filter(topic=topic)
-        if topic.poll_type == Topic.POLL_TYPE_SINGLE:
-            self.fields['answers'] = forms.ModelChoiceField(
-                label='', queryset=qs, empty_label=None,
-                widget=forms.RadioSelect())
-        elif topic.poll_type == Topic.POLL_TYPE_MULTIPLE:
-            self.fields['answers'] = forms.ModelMultipleChoiceField(
-                label='', queryset=qs,
-                widget=forms.CheckboxSelectMultiple())
-
-    def clean_answers(self):
-        answers = self.cleaned_data['answers']
-        if self.topic.poll_type == Topic.POLL_TYPE_SINGLE:
-            return [answers]
-        else:
-            return answers
