+++ pybb/forms/profile.py
@@ -1,0 +1,37 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+import re
+
+from django import forms
+from django.core.exceptions import FieldError
+from django.utils.translation import ugettext as _
+
+from pybb import defaults, util
+
+
+try:
+    class EditProfileForm(forms.ModelForm):
+        class Meta(object):
+            model = util.get_pybb_profile_model()
+            fields = ['signature', 'time_zone', 'language', 'show_signatures', 'avatar']
+
+        def __init__(self, *args, **kwargs):
+            super(EditProfileForm, self).__init__(*args, **kwargs)
+            self.fields['signature'].widget = forms.Textarea(attrs={'rows': 2, 'cols:': 60})
+
+        def clean_avatar(self):
+            if self.cleaned_data['avatar'] and (self.cleaned_data['avatar'].size > defaults.PYBB_MAX_AVATAR_SIZE):
+                forms.ValidationError(_('Avatar is too large, max size: %s bytes' %
+                                               defaults.PYBB_MAX_AVATAR_SIZE))
+            return self.cleaned_data['avatar']
+
+        def clean_signature(self):
+            value = self.cleaned_data['signature'].strip()
+            if len(re.findall(r'\n', value)) > defaults.PYBB_SIGNATURE_MAX_LINES:
+                raise forms.ValidationError('Number of lines is limited to %d' % defaults.PYBB_SIGNATURE_MAX_LINES)
+            if len(value) > defaults.PYBB_SIGNATURE_MAX_LENGTH:
+                raise forms.ValidationError('Length of signature is limited to %d' % defaults.PYBB_SIGNATURE_MAX_LENGTH)
+            return value
+except FieldError:
+    pass
