+++ pybb/models/category.py
@@ -1,0 +1,41 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.core.urlresolvers import reverse
+from django.db import models
+from django.utils.encoding import python_2_unicode_compatible
+from django.utils.translation import ugettext_lazy as _
+
+from .topic import Topic
+from .post import Post
+
+
+@python_2_unicode_compatible
+class Category(models.Model):
+    name = models.CharField(_('Name'), max_length=80)
+    position = models.IntegerField(_('Position'), blank=True, default=0)
+    hidden = models.BooleanField(_('Hidden'), blank=False, null=False, default=False,
+                                 help_text=_('If checked, this category will be visible only for staff'))
+
+    class Meta(object):
+        ordering = ['position']
+        verbose_name = _('Category')
+        verbose_name_plural = _('Categories')
+        app_label = 'pybb'
+
+    def __str__(self):
+        return self.name
+
+    def forum_count(self):
+        return self.forums.all().count()
+
+    def get_absolute_url(self):
+        return reverse('pybb:category', kwargs={'pk': self.id})
+
+    @property
+    def topics(self):
+        return Topic.objects.filter(forum__category=self).select_related()
+
+    @property
+    def posts(self):
+        return Post.objects.filter(topic__forum__category=self).select_related()
