+++ pybb/templates/pybb/avatar.html
@@ -4,18 +4,18 @@
     {% pybb_get_profile user=user as user_profile %}
     <a href="{{ user_profile.get_absolute_url }}">
         {% if user_profile.avatar %}
-            {% if_has_tag thumbnail empty %}
+            {% if_has_tag thumbnail empty %}{% comment %}sorl.thumbnail{% endcomment %}
                 {% thumbnail user_profile.avatar PYBB_AVATAR_DIMENSIONS as avatar %}
                     <img src="{{ avatar.url }}" alt="{{ user_profile.get_display_name }} avatar"/>
                 {% endthumbnail %}
             {% else %}
-                {% if_has_tag thumbnail %}
+                {% if_has_tag thumbnail %}{% comment %}easy_thumbnails{% endcomment %}
                     {% thumbnail user_profile.avatar|default:PYBB_DEFAULT_AVATAR_URL PYBB_AVATAR_DIMENSIONS as avatar %}
                     <img src="{{ avatar.url }}" alt="{{ user_profile.get_display_name }} avatar"/>
                 {% else %}
                     <img src="{{ user_profile.avatar.url }}" alt="{{ user_profile.get_display_name }} avatar" width="{{ PYBB_AVATAR_WIDTH }}" height="{{ PYBB_AVATAR_HEIGHT }}"/>
-                {% end_if_has_tag %}
-            {% end_if_has_tag %}
+                {% endif_has_tag %}
+            {% endif_has_tag %}
         {% else %}
             <img src="{{ PYBB_DEFAULT_AVATAR_URL }}" alt="default avatar" width="{{ PYBB_AVATAR_WIDTH }}" height="{{ PYBB_AVATAR_HEIGHT }}"/>
         {% endif %}
