+++ pybb/views/post.py
@@ -1,0 +1,172 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import unicode_literals
+
+import math
+
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.views import redirect_to_login
+from django.core.exceptions import PermissionDenied
+from django.core.urlresolvers import reverse
+from django.http import Http404, HttpResponse, HttpResponseRedirect
+from django.shortcuts import get_object_or_404
+from django.utils.decorators import method_decorator
+from django.views import generic
+from django.views.decorators.csrf import csrf_protect
+
+from pybb import compat, defaults, util
+from pybb.permissions import perms
+from pybb.models import Post, Forum, Topic
+
+from .mixins import RedirectToLoginMixin, PostEditMixin
+
+User = compat.get_user_model()
+username_field = compat.get_username_field()
+
+
+class PostView(RedirectToLoginMixin, generic.RedirectView):
+
+    def get_login_redirect_url(self):
+        return reverse('pybb:post', args=(self.kwargs['pk'],))
+
+    def get_redirect_url(self, **kwargs):
+        post = get_object_or_404(Post.objects.all(), pk=self.kwargs['pk'])
+        if not perms.may_view_post(self.request.user, post):
+            raise PermissionDenied
+        count = post.topic.posts.filter(created__lt=post.created).count() + 1
+        page = math.ceil(count / float(defaults.PYBB_TOPIC_PAGE_SIZE))
+        return '%s?page=%d#post-%d' % (reverse('pybb:topic', args=[post.topic.id]), page, post.id)
+
+
+class AddPostView(PostEditMixin, generic.CreateView):
+
+    template_name = 'pybb/add_post.html'
+
+    @method_decorator(csrf_protect)
+    def dispatch(self, request, *args, **kwargs):
+        if request.user.is_authenticated():
+            self.user = request.user
+        else:
+            if defaults.PYBB_ENABLE_ANONYMOUS_POST:
+                self.user, new = User.objects.get_or_create(**{username_field: defaults.PYBB_ANONYMOUS_USERNAME})
+            else:
+                return redirect_to_login(request.get_full_path())
+
+        self.forum = None
+        self.topic = None
+        if 'forum_id' in kwargs:
+            self.forum = get_object_or_404(perms.filter_forums(request.user, Forum.objects.all()), pk=kwargs['forum_id'])
+            if not perms.may_create_topic(self.user, self.forum):
+                raise PermissionDenied
+        elif 'topic_id' in kwargs:
+            self.topic = get_object_or_404(perms.filter_topics(request.user, Topic.objects.all()), pk=kwargs['topic_id'])
+            if not perms.may_create_post(self.user, self.topic):
+                raise PermissionDenied
+
+            self.quote = ''
+            if 'quote_id' in request.GET:
+                try:
+                    quote_id = int(request.GET.get('quote_id'))
+                except TypeError:
+                    raise Http404
+                else:
+                    post = get_object_or_404(Post, pk=quote_id)
+                    profile = util.get_pybb_profile(post.user)
+                    self.quote = util._get_markup_quoter(defaults.PYBB_MARKUP)(post.body, profile.get_display_name())
+
+                if self.quote and request.is_ajax():
+                    return HttpResponse(self.quote)
+        return super(AddPostView, self).dispatch(request, *args, **kwargs)
+
+    def get_form_kwargs(self):
+        ip = self.request.META.get('REMOTE_ADDR', '')
+        form_kwargs = super(AddPostView, self).get_form_kwargs()
+        form_kwargs.update(dict(topic=self.topic, forum=self.forum, user=self.user,
+                           ip=ip, initial={}))
+        if getattr(self, 'quote', None):
+            form_kwargs['initial']['body'] = self.quote
+        if perms.may_post_as_admin(self.user):
+            form_kwargs['initial']['login'] = getattr(self.user, username_field)
+        form_kwargs['may_create_poll'] = perms.may_create_poll(self.user)
+        return form_kwargs
+
+    def get_context_data(self, **kwargs):
+        ctx = super(AddPostView, self).get_context_data(**kwargs)
+        ctx['forum'] = self.forum
+        ctx['topic'] = self.topic
+        return ctx
+
+    def get_success_url(self):
+        if (not self.request.user.is_authenticated()) and defaults.PYBB_PREMODERATION:
+            return reverse('pybb:index')
+        return super(AddPostView, self).get_success_url()
+
+
+class EditPostView(PostEditMixin, generic.UpdateView):
+
+    model = Post
+
+    context_object_name = 'post'
+    template_name = 'pybb/edit_post.html'
+
+    @method_decorator(login_required)
+    @method_decorator(csrf_protect)
+    def dispatch(self, request, *args, **kwargs):
+        return super(EditPostView, self).dispatch(request, *args, **kwargs)
+
+    def get_form_kwargs(self):
+        form_kwargs = super(EditPostView, self).get_form_kwargs()
+        form_kwargs['may_create_poll'] = perms.may_create_poll(self.request.user)
+        return form_kwargs
+
+    def get_object(self, queryset=None):
+        post = super(EditPostView, self).get_object(queryset)
+        if not perms.may_edit_post(self.request.user, post):
+            raise PermissionDenied
+        return post
+
+
+class ModeratePost(generic.RedirectView):
+    def get_redirect_url(self, **kwargs):
+        post = get_object_or_404(Post, pk=self.kwargs['pk'])
+        if not perms.may_moderate_topic(self.request.user, post.topic):
+            raise PermissionDenied
+        post.on_moderation = False
+        post.save()
+        return post.get_absolute_url()
+
+
+class DeletePostView(generic.DeleteView):
+
+    template_name = 'pybb/delete_post.html'
+    context_object_name = 'post'
+
+    def get_object(self, queryset=None):
+        post = get_object_or_404(Post.objects.select_related('topic', 'topic__forum'), pk=self.kwargs['pk'])
+        if not perms.may_delete_post(self.request.user, post):
+            raise PermissionDenied
+        self.topic = post.topic
+        self.forum = post.topic.forum
+        if not perms.may_moderate_topic(self.request.user, self.topic):
+            raise PermissionDenied
+        return post
+
+    def delete(self, request, *args, **kwargs):
+        self.object = self.get_object()
+        self.object.delete()
+        redirect_url = self.get_success_url()
+        if not request.is_ajax():
+            return HttpResponseRedirect(redirect_url)
+        else:
+            return HttpResponse(redirect_url)
+
+    def get_success_url(self):
+        try:
+            Topic.objects.get(pk=self.topic.id)
+        except Topic.DoesNotExist:
+            return self.forum.get_absolute_url()
+        else:
+            if not self.request.is_ajax():
+                return self.topic.get_absolute_url()
+            else:
+                return ""
