+++ pybb/views/topic_actions.py
@@ -1,0 +1,59 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import unicode_literals
+
+from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
+from django.http import HttpResponseRedirect
+from django.shortcuts import get_object_or_404
+from django.utils.decorators import method_decorator
+from django.views import generic
+
+from pybb.permissions import perms
+from pybb.models import Topic
+
+class TopicActionBaseView(generic.View):
+
+    def get_topic(self):
+        return get_object_or_404(Topic, pk=self.kwargs['pk'])
+
+    @method_decorator(login_required)
+    def get(self, *args, **kwargs):
+        self.topic = self.get_topic()
+        self.action(self.topic)
+        return HttpResponseRedirect(self.topic.get_absolute_url())
+
+
+class StickTopicView(TopicActionBaseView):
+
+    def action(self, topic):
+        if not perms.may_stick_topic(self.request.user, topic):
+            raise PermissionDenied
+        topic.sticky = True
+        topic.save()
+
+
+class UnstickTopicView(TopicActionBaseView):
+
+    def action(self, topic):
+        if not perms.may_unstick_topic(self.request.user, topic):
+            raise PermissionDenied
+        topic.sticky = False
+        topic.save()
+
+
+class CloseTopicView(TopicActionBaseView):
+
+    def action(self, topic):
+        if not perms.may_close_topic(self.request.user, topic):
+            raise PermissionDenied
+        topic.closed = True
+        topic.save()
+
+
+class OpenTopicView(TopicActionBaseView):
+    def action(self, topic):
+        if not perms.may_open_topic(self.request.user, topic):
+            raise PermissionDenied
+        topic.closed = False
+        topic.save()
