+++ pybb/views/user.py
@@ -1,0 +1,104 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import unicode_literals
+
+from django.contrib.auth.decorators import login_required
+from django.core.urlresolvers import reverse
+from django.shortcuts import get_object_or_404
+from django.utils.decorators import method_decorator
+from django.views import generic
+from django.views.decorators.csrf import csrf_protect
+
+from pybb import compat, defaults, util
+from pybb.forms import EditProfileForm
+from pybb.permissions import perms
+from pybb.models import Topic, Post
+
+from .mixins import PaginatorMixin
+
+User = compat.get_user_model()
+username_field = compat.get_username_field()
+
+
+class UserView(generic.DetailView):
+    model = User
+    template_name = 'pybb/user.html'
+    context_object_name = 'target_user'
+
+    def get_object(self, queryset=None):
+        if queryset is None:
+            queryset = self.get_queryset()
+        return get_object_or_404(queryset, **{username_field: self.kwargs['username']})
+
+    def get_context_data(self, **kwargs):
+        ctx = super(UserView, self).get_context_data(**kwargs)
+        ctx['topic_count'] = Topic.objects.filter(user=ctx['target_user']).count()
+        return ctx
+
+
+class ProfileEditView(generic.UpdateView):
+
+    template_name = 'pybb/edit_profile.html'
+
+    def get_object(self, queryset=None):
+        return util.get_pybb_profile(self.request.user)
+
+    def get_form_class(self):
+        if not self.form_class:
+            return EditProfileForm
+        else:
+            return super(ProfileEditView, self).get_form_class()
+
+    @method_decorator(login_required)
+    @method_decorator(csrf_protect)
+    def dispatch(self, request, *args, **kwargs):
+        return super(ProfileEditView, self).dispatch(request, *args, **kwargs)
+
+    def get_success_url(self):
+        return reverse('pybb:edit_profile')
+
+
+class UserPosts(PaginatorMixin, generic.ListView):
+    model = Post
+    paginate_by = defaults.PYBB_TOPIC_PAGE_SIZE
+    template_name = 'pybb/user_posts.html'
+
+    def dispatch(self, request, *args, **kwargs):
+        username = kwargs.pop('username')
+        self.user = get_object_or_404(**{'klass': User, username_field: username})
+        return super(UserPosts, self).dispatch(request, *args, **kwargs)
+
+    def get_queryset(self):
+        qs = super(UserPosts, self).get_queryset()
+        qs = qs.filter(user=self.user)
+        qs = perms.filter_posts(self.request.user, qs).select_related('topic')
+        qs = qs.order_by('-created', '-updated', '-id')
+        return qs
+
+    def get_context_data(self, **kwargs):
+        context = super(UserPosts, self).get_context_data(**kwargs)
+        context['target_user'] = self.user
+        return context
+
+
+class UserTopics(PaginatorMixin, generic.ListView):
+    model = Topic
+    paginate_by = defaults.PYBB_FORUM_PAGE_SIZE
+    template_name = 'pybb/user_topics.html'
+
+    def dispatch(self, request, *args, **kwargs):
+        username = kwargs.pop('username')
+        self.user = get_object_or_404(User, username=username)
+        return super(UserTopics, self).dispatch(request, *args, **kwargs)
+
+    def get_queryset(self):
+        qs = super(UserTopics, self).get_queryset()
+        qs = qs.filter(user=self.user)
+        qs = perms.filter_topics(self.user, qs)
+        qs = qs.order_by('-updated', '-created', '-id')
+        return qs
+
+    def get_context_data(self, **kwargs):
+        context = super(UserTopics, self).get_context_data(**kwargs)
+        context['target_user'] = self.user
+        return context
